pyproject.toml
[tool.poetry]
name = "glue-starter-kit"
version = "0.0.1"
description = "AWS Glue Local Kit"
authors = ["Madoka Sato <msato@altus5.com>"]
# Use identifier from https://spdx.org/licenses/
license = "MIT"
readme = "README.md"
homepage = "https://github.com/msato-ok/glue-starter-kit"
repository = "https://github.com/msato-ok/glue-starter-kit"
documentation = "https://github.com/msato-ok/glue-starter-kit/blob/master/README.md"
classifiers = [
    # https://pypi.org/classifiers/
    "Development Status :: 1 - Planning",
]
[tool.poetry.urls]
# If you publish you package on PyPI, these will appear in the Project Links section.
"Bug Tracker" = "https://github.com/msato-ok/glue-starter-kit/issues"

[tool.poetry.scripts]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
pyspark = "^3.3.0"
aws-glue-libs = {path = "/opt/aws-glue-libs/"}
# AWS Glue で提供済みの Python モジュールは以下で参照する
# https://docs.aws.amazon.com/ja_jp/glue/latest/dg/aws-glue-programming-python-libraries.html

[tool.poetry.group.dev.dependencies]
bandit = "^1.6.2"
black = "^22.0"
bump2version = "^1.0.1"
isort= "^5.0.8"
flake8 = "^5.0.4"
jedi-language-server = "^0.22.0"
mypy = "^0.971"
pylint = "^2.6.0"
pytest = "^7.0.1"
pytest-cov = "^3.0"
pytest-xdist = "^2.5.0"
vulture = "^2.3"
pre-commit = "^2.20.0"
types-PyYAML = "^6.0.11"
faker = "^15.1.1"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
target-version = ['py37']

[tool.pytest.ini_options]
# Example
filterwarnings = [
    "ignore::DeprecationWarning:moto.*:",
    "ignore::DeprecationWarning:boto.*:",
]
